requirements.txt
numpy
scikit-learn
pandas
matplotlib
ravop
python-dotenv